+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Unreleased
 * Added net routines for adding and removing var generators from the network.
 * Added net routine for extracting subnetwork containing a specific set of buses.
 * Extended problem.show() to show number of vars and constraints of each type.
+* Fix bug involving problem/constraint/function's network going out of scope in Python.
 
 Version 1.3.2
 -------------